Improve Update method on ProductCategoryController

+++ EvidenceZakazekWebApp/Persistence/Repositories/PropertyDefinitionRepository.cs
@@ -1,8 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using EvidenceZakazekWebApp.Core.Models;
+using EvidenceZakazekWebApp.Core.Repositories;
+using System;
+using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
-using EvidenceZakazekWebApp.Core.Models;
-using EvidenceZakazekWebApp.Core.Repositories;
 
 namespace EvidenceZakazekWebApp.Persistence.Repositories
 {
@@ -22,11 +23,25 @@ namespace EvidenceZakazekWebApp.Persistence.Repositories
                 .ToList();
         }
 
-        public void RemoveWithValues(int id)
+        public void UpdateDefinitionsForProductCategory(int productCategoryId, ICollection<PropertyDefinition> updatedPropertyDefinitions)
+        {
+            var oldPropertyDefinitions = GetDefinitionsByCategory(productCategoryId);
+
+            var propertyDefinitionsForDelete = oldPropertyDefinitions.Where(oldPd =>
+                updatedPropertyDefinitions.Any(updatedPd =>
+                    updatedPd.Id != oldPd.Id)).ToList();
+
+            propertyDefinitionsForDelete.ForEach(pd => RemoveWithValues(pd.Id));
+        }
+
+        private void RemoveWithValues(int id)
         {
             var definitionForRemove = _context.PropertyDefinitions
-                        .Include(pd => pd.PropertyValues)
-                        .SingleOrDefault(pd => pd.Id == id);
+                .Include(pd => pd.PropertyValues)
+                .SingleOrDefault(pd => pd.Id == id);
+
+            if (definitionForRemove == null)
+                throw new ArgumentNullException(nameof(definitionForRemove));
 
             _context.PropertyValues.RemoveRange(definitionForRemove.PropertyValues);
             _context.PropertyDefinitions.Remove(definitionForRemove);
